# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.CreateNoteRequest:
    properties:
      content:
        type: string
      title:
        type: string
    type: object
  models.Note:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      title:
        type: string
    required:
    - content
    - title
    type: object
  models.UpdateNoteRequest:
    properties:
      content:
        type: string
      title:
        type: string
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:8080
info:
  contact: {}
  description: This is a sample server celler server.
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /notes:
    get:
      consumes:
      - application/json
      description: get notes
      produces:
      - application/json
      responses:
        "200":
          description: Список заметок
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties: true
            type: object
      summary: List notes
      tags:
      - notes
    post:
      consumes:
      - application/json
      description: creating a note
      parameters:
      - description: Данные заметки
        in: body
        name: note
        required: true
        schema:
          $ref: '#/definitions/models.CreateNoteRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Заметка создана
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Неверные данные
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties: true
            type: object
      summary: Create a note
      tags:
      - notes
  /notes/{id}:
    delete:
      consumes:
      - application/json
      description: delete note by ID
      parameters:
      - description: ID заметки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Заметка удалена
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Неверный ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Заметка не найдена
          schema:
            additionalProperties: true
            type: object
      summary: Delete note
      tags:
      - notes
    get:
      consumes:
      - application/json
      description: get note by ID
      parameters:
      - description: ID заметки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Заметка найдена
          schema:
            $ref: '#/definitions/models.Note'
        "400":
          description: Неверный ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Заметка не найдена
          schema:
            additionalProperties: true
            type: object
      summary: Show note
      tags:
      - notes
    put:
      consumes:
      - application/json
      description: update note by ID
      parameters:
      - description: ID заметки
        in: path
        name: id
        required: true
        type: integer
      - description: Данные для обновления
        in: body
        name: updateData
        required: true
        schema:
          $ref: '#/definitions/models.UpdateNoteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Заметка обновлена
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Неверные данные
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Заметка не найдена
          schema:
            additionalProperties: true
            type: object
      summary: Update note
      tags:
      - notes
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"